==> scripts/00_data-summaries.R <==
# Create tables summarizing the two types of data used

# For a pub I would like to see a table that summarize the basic data. 
#   - Number of shots per survey
#   - How many with positive data
#   - How many sampled, etc. 

# Do a separate table for Atlantic herring

# Use formatted data or raw data?

library("tidyverse")
library("here")


# Load helper functions
source(here("functions", "process_trawl_data.R"))
source(here("functions", "process_diet_data.R"))

# Raw data ----------------------------------------------------------------
dietraw_0 <- readr::read_rds(here("data", "processed", "dat_preds_all.rds")) 

# What percent of total stomach contents observed is from my 5 predators?
ms <- dietraw_0 %>% mutate(my_pred = ifelse(predator %in% c("atlantic cod", "silver hake", "spiny dogfish", "goosefish", "white hake"),
                                  TRUE, FALSE)) %>%
  group_by(my_pred) %>%
  summarize(tot_py_w = sum(pyamtw, na.rm = FALSE),
            n_preds = n())
# weight
347570/(347570 + 47674)
# number
177937/(111533 + 177937)

dietraw <- dietraw_0 %>%
  dplyr::filter(predator %in% c("atlantic cod", "silver hake", "spiny dogfish", "goosefish", "white hake"))

trawlraw <- readr::read_rds(here("data", "processed", "dat_trawl.rds"))


# Mean dates for spring and fall surveys ----------------------------------
survey_dates <- dietraw %>%
  select(towid, month, day, year, season) %>%
  distinct() %>%
  drop_na() %>%
  mutate(date = lubridate::ymd(paste(2000, month, day))) # need to use a leap year

ggplot(survey_dates, aes(x = date, fill = season)) +
  geom_bar() +
  geom_vline(xintercept = lubridate::ymd("2000-09-01"))

# How many samples in "fall" before September?
survey_dates %>%
  filter(season == "fall") %>%
  summarize(
    n_summer = sum(date < lubridate::ymd("2000-09-01")),
    n = n()
  )
100*654/8527

md <- survey_dates %>%
  group_by(season) %>%
  summarize(mean = mean(date),
            median = median(date),
            min = min(date),
            max = max(date))

# Processed data ----------------------------------------------------------
dietsetup <- readr::read_rds(here("data", "processed", "dat_preds_all.rds")) %>%
  dplyr::filter(predator %in% c("atlantic cod", "silver hake", "spiny dogfish", "goosefish", "white hake")) %>%
  group_by(predator, season) %>%
  nest() %>%
  mutate(processed_data = purrr::map(data, process_diet_data))

trawlsetup <- readr::read_rds(here("data", "processed", "dat_trawl.rds")) %>%
  group_by(species, season) %>%
  nest() %>%
  mutate(processed_data = purrr::map(data, process_trawl_data))





# Check vessels -----------------------------------------------------------

group_by(trawlraw, vessel, year, season) %>% summarize(n = n()) %>%
  pivot_wider(names_from = "vessel", values_from = "n")

mutate(trawlraw, boat_season = paste(season, vessel)) %>%
  group_by(boat_season, year) %>% 
  summarize(n = n()) %>%
  pivot_wider(names_from = "boat_season", values_from = "n")

# Try to link observations using date/location metadata as a stand in for vessel
dietobs <- dplyr::filter(dietraw, towid == "197303 12")
trawlobs <- dplyr::filter(trawlraw, towid == dietobs$towid & species == dietobs$predator)

# Based on location, this is from DE
# which makes sense because AT was only used for a few inshore surveys
# could look through diet data and see if, based on location, there's evidence of any
# diet samples from AT
filter(trawlobs, declat == dietobs$declat)



# First glance using good data ---------------------------------------------
# Only need to worry about periods when more than one vessel was being used:
#   - spring and fall 1973-1981 (except spring 1978)
#   - fall 1985
#   - spring 1987
#   - fall 1988
#   - spring 1992
# So reate a flag in raw diet data to separate these bad ones

num_boats <- group_by(trawlraw, vessel, year, season) %>% 
  summarize(n = n()) %>%
  filter(n > 0) %>%
  group_by(year, season) %>%
  summarize(n_boats = n()) %>%
  ungroup() 

# Add number of boats to trawl data
trawl_boat <- full_join(trawlraw, num_boats, by = c("year", "season"))

# Add number of boats to diet data
anti_join(dietraw, num_boats, by = c("year", "season"))
diet_boat <- full_join(dietraw, num_boats, by = c("year", "season"))




# Another approach: widen diet data to one observation per row (essentially what Jakub sent)
# Drop unneccessary data to make it easier (keep lengths?)

# Tows should all have the same values for covariates...
trawlunique <- filter(trawl_boat, n_boats == 1) %>%
  filter(species != "atlantic herring") %>%
  rename(predator = species)

# check that ids are unique between seasons
# do any towids show up in both seasons?
trawlunique %>% 
  select(towid, year, season) %>%
  distinct() %>%
  pull(towid) %>%
  duplicated() %>%
  sum()

dietunique <- filter(diet_boat, n_boats == 1) %>%
  filter(predator %in% c("atlantic cod", "silver hake", "spiny dogfish", "goosefish", "white hake")) %>%
  group_by(predator, season, towid) %>%
  summarize(
    # declat_var = var(declat),
    # declon_var = var(declon),
    declat_avg = mean(declat, na.rm = TRUE),
    declon_avg = mean(declon, na.rm = TRUE),
    n_pred = n(),
    length_avg = mean(pdlen),
    pdwgt_avg = mean(pdwgt),
    pyamtw_tot = sum(pyamtw),
    pypres_tot = sum(pypres)
  )

# Check that locations are consistent
# sum(dietunique$declat_var > 0, na.rm = TRUE)
# sum(dietunique$declon_var > 0, na.rm = TRUE)

# Join by towid (I've already subsetted the unique tows)
# and by predator/species

anti_join(trawlunique, dietunique, by = c("towid", "predator")) # Expect that many tows won't have diet data
anti_join(dietunique, trawlunique, by = c("towid", "predator")) # but this is more worrying, that these don't match

alldat <- full_join()

# So I can continue to try and work around this, but I think the smarter thing
# is to do a data request from Jakub and clarify what the tow ID is here

# What do I need/what questions do I have?
# 1) do another pull of this diet data query, including the vessel
# 2) what combination of columns will give me unique tows? 
#     - should be cruise6 + station
#     - why for trawl data did cruise6 + station + vessel not give unique? just one, and there is also a tow number that probably makes it unique


# hmmm if cruise6 + station (aka towid) is NOT unique (i.e. it also needs vessel)
# then I should check diet data and see if there is any variance in location,
# month, etc. to see if it is actually combining across multiple tows
jj <- diet_boat %>%
  filter(predator %in% c("atlantic cod", "silver hake", "spiny dogfish", "goosefish", "white hake")) %>%
  group_by(towid) %>%
  summarize(
    season = n_distinct(season),
    station = n_distinct(station),
    geoarea = n_distinct(geoarea),
    declat_var = n_distinct(declat),
    declon_var = n_distinct(declon),
    month = n_distinct(month),
    day = n_distinct(day),
    setdepth = n_distinct(setdepth)
    ) %>%
  pivot_longer(cols = -towid, names_to = "column", values_to = "value")

filter(jj, value > 1)


# Try again with even more raw data
frdata <- read_csv(here::here("data", "raw", "fr_diet.csv"), guess_max = 365080) %>%
  dplyr::select(-X1) 

myfrdata <- frdata %>% 
  mutate(
    towid = paste(cruise6, station)
  ) %>%
  group_by(towid) %>%
  summarize(
    season = n_distinct(season),
    station = n_distinct(station),
    geoarea = n_distinct(geoarea),
    declat_var = n_distinct(declat),
    declon_var = n_distinct(declon),
    month = n_distinct(month),
    day = n_distinct(day),
    setdepth = n_distinct(setdepth)
  ) %>%
  pivot_longer(cols = -towid, names_to = "column", values_to = "value")

filter(myfrdata, value > 1)


# The rawest data join possible -------------------------------------------
# Also import everything as text, but maybe still leading zeros have gotten dropped!!
# Try with *even more* raw format of trawl data
nohakedat <- read_csv(here("data", "raw", "Ng_OPS.txt"), # guess_max = 365080,
                      col_types = cols(.default = "c"))
trawldat <- nohakedat
names(trawldat) <- tolower(names(nohakedat))

# Let's just look at individual tows from frdata
dietdat <- read_csv(here::here("data", "raw", "fr_diet.csv"), col_types = cols(.default = "c")) %>%
  select(cruise6, station, stratum, year, season, declat, declon, setdepth) %>%
  distinct()

dietdat_fix <- mutate(dietdat, nzeros = 4 - nchar(station)) %>%
  mutate(station = pmap_chr(list(nzeros, station), 
                             function(a,b) paste0(paste0(rep(0, a), collapse = ""), b, collapse = "")))


# tows in trawl data that don't match diet data (expected) 10091 rows
anti_join(trawldat, dietdat_fix, by = c("cruise6", "station")) %>% nrow() 

# tows in diet data that don't match trawl data (not expected) 674 rows
anti_join(dietdat_fix, trawldat, by = c("cruise6", "station")) %>% nrow()

# tows in diet data that do match trawl data 24193
semi_join(dietdat_fix, trawldat, by = c("cruise6", "station")) %>% nrow()

# out of total possible rows in diet data: 24867
nrow(dietdat_fix)

# How many rows have multiple mathches?
left_join(dietdat_fix, trawldat, by = c("cruise6", "station"))
# Adding stratum will probably resolve these (easier to get equal than location)

# I think I need to pad stratum too
summary(nchar(trawldat$stratum))
summary(nchar(dietdat_fix$stratum))

dietdat3 <- mutate(dietdat_fix, nzeros = 5 - nchar(stratum)) %>%
  mutate(stratum = pmap_chr(list(nzeros, stratum), 
                            function(a,b) paste0(paste0(rep(0, a), collapse = ""), b, collapse = "")))
# Ok check again
# tows in trawl data that don't match diet data (expected) 10091 rows -> 10266 hmm (increase of 175), but multiple matches potentially??
anti_join(trawldat, dietdat3, by = c("cruise6", "station", "stratum")) %>% nrow() 

# tows in diet data that don't match trawl data (not expected) 674 rows -> 683 (increase of 9)
anti_join(dietdat3, trawldat, by = c("cruise6", "station", "stratum")) %>% nrow()

# tows in diet data that do match trawl data 24193 -> 24184 (decrease of 9)
semi_join(dietdat3, trawldat, by = c("cruise6", "station", "stratum")) %>% nrow()

# How many rows have multiple mathches?
tmp <- left_join(dietdat3, trawldat, by = c("cruise6", "station", "stratum"))

filter(tmp, is.na(year.y)) # 683 diet obs with no match in trawl data

# Ok so where am I? I can join the diet data by cruise6, station, and stratum and get no duplicates
# (i.e., I think this accoutns for the duplicated vessel station codes)
# BUT I am missing trawl data for some cruises. 

filter(tmp, setdepth!=depth | is.na(depth))

# Write a csv of the diet observations that are missing corresponding trawl information
# then email Jakub to get more info. 
notrawls <- anti_join(dietdat3, trawldat, by = c("cruise6", "station", "stratum"))
semi_join(tmp, notrawls, by = c("cruise6", "station" ,"stratum")) %>% View()

# And quickly, how many diet data observations are therefore missing trawl data?
all_diet <- read_csv(here::here("data", "raw", "fr_diet.csv"), col_types = cols(.default = "c")) %>%
  mutate(nzeros = 4 - nchar(station)) %>%
  mutate(station = pmap_chr(list(nzeros, station), 
                            function(a,b) paste0(paste0(rep(0, a), collapse = ""), b, collapse = ""))) %>%
  mutate(nzeros = 5 - nchar(stratum)) %>%
  mutate(stratum = pmap_chr(list(nzeros, stratum), 
                            function(a,b) paste0(paste0(rep(0, a), collapse = ""), b, collapse = "")))
semi_join(all_diet, notrawls, by = c("cruise6", "station", "stratum"))
# 15226/365079 rows, so about 4% (can probably proceed with table as is while I wait for clarification)

# But first, pull data with all trawl-level data, so drop predator data from this last join
all_diet_distinct <- all_diet %>%
  select(cruise6, station, stratum, setdepth, beglat, beglon, declat, declon, month, day,
         year, purcode, season, geoarea) %>%
  distinct() 

check <- all_diet_distinct %>%
  left_join(notrawls, by = c("cruise6", "station", "stratum"))

# Hypothesis, maybe missing trawls by region?
filter(check, is.na(year.y)) %>%
  pull(geoarea) %>%
  as.factor() %>%
  summary()

filter(check, !is.na(year.y)) %>%
  pull(geoarea) %>%
  as.factor() %>%
  summary()
# Nope, sort of all over. Ok I'm wasting time, just send these data to Jakub and ask.
all_diet_no_trawls <- semi_join(all_diet_distinct, notrawls, by = c("cruise6", "station", "stratum"))
write_csv(all_diet_no_trawls, path = here("data", "processed", "diets_missing_trawl.csv"))

################################################################################
# Do the quick and dirty data summarization -------------------------------
# Instead of dropping rows, pivot wider like trawldat? Or work with rows and then join later
# Quantities that I actually want in final table: 
#     - number tows per year-season
#     - percent tows with each predator present
#     - given predators present, what percent had herring?

library("tidyverse")
library("here")

# Load trawl data as all character data
nohakedat <- read_csv(here("data", "raw", "Ng_OPS.txt"), col_types = cols(.default = "c"))
withhake <- read_csv(here("data", "raw", "Ng_OPS_urophycis_tenuis.txt"), col_types = cols(.default = "c"))

rawtrawl <- bind_cols(nohakedat, withhake[, 16]) 

trawldat <- rawtrawl %>%
  filter(!(as.numeric(LAT) > 37.5 & as.numeric(LON) < -75.6)) %>%
  filter(!(as.numeric(LAT) < 34.5 & as.numeric(LON) > -75.5)) %>%
  rename(
    herring_kg = CLUPEA_HARENGUS_WT,
    ATLANTIC_COD_wt = GADUS_MORHUA_WT,
    SILVER_HAKE_wt = MERLUCCIUS_BILINEARIS_WT,
    SPINY_DOGFISH_wt = SQUALUS_ACANTHIAS_WT,
    GOOSEFISH_wt = LOPHIUS_AMERICANUS_WT,
    WHITE_HAKE_wt = UROPHYCIS_TENUIS_WT
  ) %>%
  pivot_longer(
    cols = contains("_wt"),
    names_to = "predator",
    values_to = "catch_kg"
  ) %>%
  mutate(
    predator = str_remove(predator, "_wt"),
    predator = str_replace(predator, "_", " "),
    predator = tolower(predator)
  ) %>%
  mutate(
    catch_kg = ifelse(catch_kg == 0, 9999999999, catch_kg),
    catch_kg = replace_na(catch_kg, 0),
    catch_kg = ifelse(catch_kg == 9999999999, "present", catch_kg)
  ) %>%
  mutate(
    herring_kg = ifelse(herring_kg == 0, 9999999999, herring_kg),
    herring_kg = replace_na(herring_kg, 0),
    herring_kg = ifelse(herring_kg == 9999999999, "present", herring_kg)
  ) %>%
rename_all(tolower)

# Maybe don't need to group by year?
trawl_summary <- trawldat %>%
  mutate(herring_present = herring_kg !=0,
         predator_present = catch_kg !=0,
         both_present = ifelse(herring_present & predator_present, TRUE, FALSE)) %>%
  mutate(season = ifelse(season == "FALL"|season == "SUMMER", "FALL", "SPRING")) %>%
  group_by(year, season, predator) %>%
  summarize(n_tows = n(),
            pred_present = sum(predator_present)/n_tows,
            both_present = sum(both_present)/n_tows) %>%
  ungroup() %>%
  mutate(season = str_to_title(season),
         predator = str_to_sentence(predator))
  
# Load diet data as all character data and fix station and stratum codes
rawdiet <- read_csv(here::here("data", "raw", "fr_diet.csv"), col_types = cols(.default = "c"))

dietdat <- select(rawdiet, -X1) %>%
  mutate(nzeros = 4 - nchar(station)) %>%
  mutate(station = pmap_chr(list(nzeros, station), 
                            function(a,b) paste0(paste0(rep(0, a), collapse = ""), b, collapse = ""))) %>%
  mutate(nzeros = 5 - nchar(stratum)) %>%
  mutate(stratum = pmap_chr(list(nzeros, stratum), 
                            function(a,b) paste0(paste0(rep(0, a), collapse = ""), b, collapse = ""))) %>%
  select(-nzeros) %>%
  mutate(season = ifelse(season == "FALL"|season == "SUMMER", "FALL", "SPRING")) %>%
  mutate(predator = str_to_sentence(pdcomnam))%>%
  filter(predator %in% c("Atlantic cod", "Silver hake", "Spiny dogfish", "Goosefish", "White hake"))


# Get number predator stomachs per tow by species and percent with prey present
# I think I should average by tow, since that's my unit of inference
# At end, get total number of stomachs and percent with positive prey
# I can get more guidance if I need it.
diet_summary <- dietdat %>%
  mutate(prey_present = pyamtw != 0) %>%
  group_by(cruise6, station, stratum, setdepth, year, season, predator) %>%
  summarize(n_stomachs = n(),
            n_w_prey = sum(prey_present)) %>% # per tow! could skip this, but keep incase I need it?
  group_by(year, season, predator) %>%
  summarize(n_diet_tows = n(),
            n_stomachs = sum(n_stomachs),
            prey_present = sum(n_w_prey)/n_stomachs) %>%
  ungroup() %>%
  mutate(season = str_to_sentence(season))

# Join trawl data with diet data for final table
full_summary <- full_join(trawl_summary, diet_summary, by = c("predator", "year", "season")) %>%
  filter(!is.na(year)) %>% 
  replace_na(list(n_stomachs = 0, prey_present = 0))

write_csv(full_summary, path = here("output", "data-summary.csv"))

# Make an even smaller summary table
small_summary <- full_summary %>% 
  replace_na(list(n_stomachs = 0, prey_present = 0)) %>%
  group_by(predator, season) %>%
  summarize(n_tows = sum(n_tows),
            pred_pres_avg = mean(pred_present),
            both_pres_avg = mean(both_present),
            n_stomachs = sum(n_stomachs),
            prey_pres_avg = mean(prey_present))
write_csv(small_summary, path = here("output", "data-summary-small.csv"))


# Just realized I used all prey data, just look at my processed data to do these calcualtions
dietsetup <- readr::read_rds(here("data", "processed", "dat_preds_all.rds")) %>%
  dplyr::filter(predator %in% c("atlantic cod", "silver hake", "spiny dogfish", "goosefish", "white hake"))

diet_summary <- dietsetup %>%
  group_by(towid, year, season, predator) %>%
  summarize(n_stomachs = n(),
            n_w_prey = sum(pypres)) %>% # per tow! could skip this, but keep incase I need it?
  group_by(year, season, predator) %>%
  summarize(n_diet_tows = n(),
            n_stomachs = sum(n_stomachs),
            prey_present = sum(n_w_prey)/n_stomachs) %>%
  ungroup() %>%
  mutate(season = str_to_sentence(season),
         predator = str_to_sentence(predator),
         year = as.character(year))

# Join trawl data with diet data for final table
full_summary <- full_join(trawl_summary, diet_summary, by = c("predator", "year", "season")) %>%
  replace_na(list(n_diet_tows = 0, n_stomachs = 0, prey_present = 0)) %>%
  select(predator, season, year, contains("_")) %>%
  arrange(
    predator,
    desc(season),
    year)

write_csv(full_summary, path = here("output", "data-summary.csv"))
