==> scripts/00_process_consumption_data.Rmd <==
---
title: "Process consumption data for catch standardization"
output: html_notebook
editor_options: 
  chunk_output_type: inline
---

# Plan
1. Create a data frame where each row is a sampled predator. Include predator specific information (weight, species), tow specific information (towid, season, location, depth, year). This approach doesn't account for the correlation of predators within tows.
2. Create a data frame where each row is a unique tow. For the average tow level consumption, have a column for average consumption. Include average predator length and sample size. 

Later: I can incorporate other tow covariates, like amount of prey or predator in tow
4. Include stomach volume or mass for total

# Preliminaries
```{r setup}
library("tidyverse")
library("here")

save_data <- TRUE
use_all_clupea <- TRUE   # if TRUE use all clupea spp. otherwise use only C. harengus
use_empty_stomachs <- TRUE # include predators whose stomachs were empty?
```

```{r load-raw-diet-data}
frdata <- read_csv(here::here("data", "raw", "fr_diet.csv"), guess_max = 365080) %>%
  dplyr::select(-X1)
```

# 0. Clean data
Make longitude negative, remove 7 observations missing date and location data, and one spatial outlier
```{r fix-locations}
frdata <- frdata %>% 
  mutate(declon  = -declon) %>%
  filter(!is.na(month)) %>%
  filter(declon > -78)
```


Unique tow ID is cruise6 + station
Unique predator ID needs to account for cruise, station, species, replicate, and sex
```{r create-unique-IDs}
frdata <- frdata %>% 
  mutate(
    towid = paste(cruise6, station),
    predid = paste(cruise6, station, svspp, pdid, pdsex)
  )
```


# 1. Create a data frame where each row is a sampled predator
Get the ids and predator data, match it up with consumption

## Prey Data
I will not use eggs, larvae, or scales, but like Deroba 2019, I will use the Clupidae category, since these are the most common clupeidae. All prey weights are real numbers. 
```{r check-herring-weights}
filter(frdata, analcat == "CLUFAM") %>%
  group_by(pynam) %>%
  summarize(mean = mean(pyamtw),
            sd = sd(pyamtw),
            n = n())

slice(frdata, frdata$pynam %>% str_which("CLUP")) %>%
  group_by(pynam) %>%
  summarize(mean = mean(pyamtw),
            sd = sd(pyamtw),
            n = n())
```

```{r filter-out-prey}
if(use_all_clupea == TRUE){
  preydat <- frdata %>% 
    filter(pynam == "CLUPEA HARENGUS" | pynam == "CLUPEIDAE") %>%
    dplyr::select(predid, pyamtw)
}else if(use_all_clupea == FALSE){
  preydat <- frdata %>% 
    filter(pynam == "CLUPEA HARENGUS") %>%
    dplyr::select(predid, pyamtw)
}
```

Check that there is just one entry per predator, and also look at the weights. Note there is one entry with a zero weight, but that won't matter once I merge with the predator data anyway.
```{r check-prey-data}
sum(duplicated(preydat$predid)) > 0 # any duplicates?

summary(preydat$pyamtw) # any zero weights?
sum(preydat$pyamtw == 0)
```

## Predator Data
Create a data frame with one row per predator and all the relevant predator/tow info. 

Version (1) Exclude predators whose stomachs were empty, because we don't know why they were empty (e.g., regurgitated in trawl due to barotrauma).

Version (2) Empty stomachs are data, include those too. 
```{r simplify-predator-data}
if(use_empty_stomachs == FALSE){
  preddat <- frdata %>%
    filter(pynam != "EMPTY") %>% # remove empty stomachs, from 289,470 to 162,203
    dplyr::select(towid, station, geoarea, declat, declon, month, day, year, season, setdepth,
                  predid, pdcomnam, pdsex, pdlen, sizecat, pdwgt) %>%
    distinct()
}else if(use_empty_stomachs == TRUE){
  preddat <- frdata %>%
    dplyr::select(towid, station, geoarea, declat, declon, month, day, year, season, setdepth,
                  predid, pdcomnam, pdsex, pdlen, sizecat, pdwgt) %>%
    distinct()
}

sum(duplicated(preddat$predid)) > 0 # any duplicates?
```

## Merge predators and prey
Now join with data frame with the rest of the predator data. It should have the same number of rows as predid unless I have wonky duplicated stuff that needs more investigation. Need to convert NAs to 0s for unobserved prey. 
```{r merge-predator-and-consumption-data}
dat_preds_all <- left_join(preddat, preydat, by = "predid") %>%
  mutate(pyamtw = ifelse(is.na(pyamtw), 0, pyamtw),
         pypres = ifelse(pyamtw == 0, 0, 1))
```


# 2. Create a data frame where each row is a unique tow. 
Need to take the data frame I have now and condense by tow/species combo (averaging within tows).
```{r pool-predators-in-tows}
pooled_tows <- dat_preds_all %>% 
  group_by(towid, pdcomnam) %>%
  summarize(mean_pyamtw = mean(pyamtw), 
            mean_pdlen = mean(pdlen),
            tot_pypres = sum(pypres),
            n_preds = n())
```

Add back in the tow-level information to the pooled predator data by merging with tibble of tow info.  
```{r get-one-tow-per-row}
towdat <- frdata %>% 
  dplyr::select(towid, station, geoarea, declat, declon, month, day, year, season, setdepth) %>%
  distinct()

sum(duplicated(preddat$predid)) > 0 # any duplicates?

dat_tows_all <- right_join(towdat, pooled_tows, by = "towid")
```
I had previously then joined with all tow data, but here, I'm just looking at tows with predators that had non-empty stomachs. We don't want to look at tows that had no predators, or tows that had predators with no consumption whatsoever. 


# 3. Condense season names
Condense the four season designation by NEFSC into 2 seasons. The groupings are Fall and Summer (FALL) and Winter and Spring (spring), based on which seasons were closest together.
```{r fix-seasons}
dat_preds_all <- mutate(dat_preds_all, myseason = ifelse(season == "FALL"|season == "SUMMER", "FALL", "SPRING"))
dat_tows_all <- mutate(dat_tows_all, myseason = ifelse(season == "FALL"|season == "SUMMER", "FALL", "SPRING"))
```


# 4. Rename and reformat outputs
```{r format-data}
dat_preds_all <- dat_preds_all %>%
  select(-season) %>%
  rename(
    season = myseason,
    predator = pdcomnam
  ) %>%
  mutate(
    season = tolower(season),
    predator = tolower(predator)
  )

dat_tows_all <- dat_tows_all %>%
  select(-season) %>%
  rename(
    season = myseason,
    predator = pdcomnam
  ) %>%
  mutate(
    season = tolower(season),
    predator = tolower(predator)
  )
```


# 4. Save output
```{r save-output}
if(save_data == TRUE){
  saveRDS(dat_preds_all, here::here("data", "processed", "dat_preds_all.rds"))
  saveRDS(dat_tows_all, here::here("data", "processed", "dat_tows_all.rds"))
}
```


*dat_preds_all has one row for each predator with non-empty stomach (it does not include tows where predators did not occur or had empty stomachs)
*dat_tows_all has one row for each species per tow. predator info is averaged accross predators of the same species within the same tow. It does not contain information about tows that did not have any predators in it (i.e., rows where n_preds == 0), or where tows had predators with only empty stomachs.
